fix(rl): harden the RL path for the non-NVIDIA platforms
Six defects surfaced by RL E2E on the newly verified backends
(cambricon/metax/hygon/enflame/mthreads), each blocking `train_rl` on at
least one of them:

- arguments.py registers --return-log-probs unconditionally, and train_rl
  passes add_inference_args as extra_args_provider, so the flag was
  registered twice and argparse aborted with "conflicting option string"
  before training ever started (metax, hygon).
- MUSA torch ships an importable torch.cuda.nvtx stub that raises at call
  time, so the hasattr guard in PlatformMUSA passed and the first rollout
  crashed with "NVTX functions not installed". Probe a real call once and
  cache the result.
- CNNL rejects int32 index tensors in advanced indexing: extract_log_probs
  used an int32 seq_idx (BAD_PARAM on log_probs[seq_idx, ...]), and the
  resume path indexed the int32 request_kv_block_counts with an int64 row
  index (cambricon). Align both dtypes.
- torch.Generator(device=torch.cuda.current_device()) is not portable:
  enflame GCU rejects it. Sample with a CPU generator, which torch
  multinomial requires anyway for non-CPU probability tensors.
- Dynamic-batching decode pads the query batch to TOKEN_ROUNDER rows while
  block_table/seqlens_k only cover the active requests; flash-attn enforces
  block_table rows == q batch dim, so trim the padded rows for the kernel
  call and pad the output back (enflame).
- The [rl] extra lacked einops, which the dynamic-batching attention path
  imports (rearrange).

Verified on the five backends: TRAIN_RL_EXIT=0 with the eval metrics
passing on both the flagtree and triton compiler paths.

+++ megatron/core/inference/contexts/dynamic_context.py
@@ -2316,7 +2316,9 @@ class DynamicInferenceContext(BaseInferenceContext):
 
                 # Apply updates only to the requests that required a new block
                 relative_row_idx = torch.nonzero(needs_new_block).squeeze(1)
-                row_idx = resume_start + relative_row_idx
+                # request_kv_block_counts is int32; advanced indexing into it with
+                # an int64 index tensor raises CNNL BAD_PARAM, so align the dtype.
+                row_idx = (resume_start + relative_row_idx).to(self.request_kv_block_counts.dtype)
                 col_idx = self.request_kv_block_counts[row_idx]
 
                 self.request_to_kv_block_ids[row_idx, col_idx] = block_ids
@@ -2919,7 +2921,9 @@ class DynamicInferenceContext(BaseInferenceContext):
         logits_squeezed = logits.squeeze(0).float()
 
         if only_last_token_logits or self.is_decode_only():
-            seq_idx = torch.arange(len(new_tokens), dtype=torch.int32, device=logits.device)
+            # seq_idx must be int64: CNNL advanced indexing (log_probs[seq_idx, ...])
+            # rejects int32 indices with BAD_PARAM.
+            seq_idx = torch.arange(len(new_tokens), device=logits.device)
             log_probs = F.log_softmax(logits_squeezed[seq_idx], dim=-1)
             selected_log_probs = log_probs[seq_idx, new_tokens]
             return [[lp] for lp in selected_log_probs.tolist()], log_probs
